project 11

+++ src/utils/api.js
@@ -24,6 +24,7 @@ class Api extends React.Component {
   };
 
   createCard = (cardData) => {
+    console.log(cardData);
     return fetch(`${this.props.baseUrl}/cards`, {
       method: 'POST',
       headers: this.props.headers,
@@ -31,18 +32,18 @@ class Api extends React.Component {
     }).then((res) => this._getResponseData(res));
   };
 
-  addLike = (cardId) => {
-    return fetch(`${this.props.baseUrl}/cards/likes/${cardId}`, {
-      method: 'PUT',
-      headers: this.props.headers,
-    }).then((res) => this._getResponseData(res));
-  };
-
-  deleteLike = (cardId) => {
-    return fetch(`${this.props.baseUrl}/cards/likes/${cardId}`, {
-      method: 'DELETE',
-      headers: this.props.headers,
-    }).then((res) => this._getResponseData(res));
+  changeLikeCardStatus = (cardId, isLiked) => {
+    if (isLiked) {
+      return fetch(`${this.props.baseUrl}/cards/likes/${cardId}`, {
+        method: 'PUT',
+        headers: this.props.headers,
+      }).then((res) => this._getResponseData(res));
+    } else {
+      return fetch(`${this.props.baseUrl}/cards/likes/${cardId}`, {
+        method: 'DELETE',
+        headers: this.props.headers,
+      }).then((res) => this._getResponseData(res));
+    }
   };
 
   deleteCard = (cardId) => {
@@ -60,11 +61,11 @@ class Api extends React.Component {
     }).then((res) => this._getResponseData(res));
   };
 
-  updateProfilePic = (userData) => {
+  updateAvatar = (avatar) => {
     return fetch(`${this.props.baseUrl}/users/me/avatar`, {
       method: 'PATCH',
       headers: this.props.headers,
-      body: JSON.stringify(userData),
+      body: JSON.stringify(avatar),
     }).then((res) => this._getResponseData(res));
   };
 };
